Make some optimizations to clipmap_events
The most important one is stopping early the min_enter_exit_lod.
Branching and bounding to find the chunks on the clip sphere surface
turns out to be very expensive.

I even tried replacing recursion with a manual stack-based algorithm,
but it made no difference, so that approach was discarded.