Fix a mesh overlap bug
We used to have a SyncBatch<LodChange3> to allow parallelism between the
level_of_detail system and the mesh_generator system. This was flawed
because it allowed for multiple frames of commands to be processed by
the mesh generator in a single frame. The mesh generator does not
support this, since it will first apply all mesh removals and then all
mesh spawns. This would be out of order for the case where a mesh is
deleted in one frame and then created again in the next frame.